Entry refactor attempt 1

+++ src/data/modifier/TicketType.java
@@ -52,7 +52,7 @@ public class TicketType implements Serializable, Modifier, WizardType {
         return name;
     }
 
-    public long getPrice() {
+    public int getPrice() {
         return price;
     }
